Set single source of truth for prices and ensure that keyboard action does not cause border to appear in Checkout page

+++ CheckoutKata.UI/Models/SkuCard.cs
@@ -1,3 +1,3 @@
 ﻿namespace CheckoutKata.UI.Models;
 
-public record SkuCard(string Code, int Price, SkuOffer? Offer);
+public record SkuCard(string Code, decimal Price, SkuOffer? Offer);
